Add data buffer and calibration matrix processing

+++ biosiglive/processing/data_processing.py
@@ -10,11 +10,13 @@ from typing import Union
 
 try:
     from pyomeca import Analogs
+
     pyomeca_module = True
 except ModuleNotFoundError:
     pyomeca_module = False
 
 
+# TODO add a calibrate fucntion for calibration matrix
 class GenericProcessing:
     def __init__(self):
         """
@@ -25,7 +27,8 @@ class GenericProcessing:
         self.lpf_lcut = 5
         self.lp_butter_order = 4
         self.bp_butter_order = 4
-        self.ma_win = 200
+        self.moving_average_windows = 200
+        self.data_rate = 2000
 
     @staticmethod
     def _butter_bandpass(lowcut, highcut, fs, order=5):
@@ -91,51 +94,124 @@ class GenericProcessing:
             norm_emg[emg, :] = emg_data[emg, :] / mvc_list[emg]
         return norm_emg
 
+    @staticmethod
+    def calibration_matrix(data: np.ndarray, matrix: np.ndarray):
+        """
+        Apply a calibration matrix to the data.
+        Parameters
+        ----------
+        data: numpy.ndarray
+            Data to calibrate.
+        matrix: numpy.ndarray
+            Calibration matrix.
+
+        Returns
+        -------
+
+        """
+        return np.dot(data, matrix)
+
+    def _process_emg(self, data: np.ndarray,
+                    mvc_list: Union[list, tuple] = None,
+                    band_pass_filter=True,
+                    low_pass_filter=False,
+                    moving_average=True,
+                    centering=True,
+                    absolute_value=True,
+                    normalization=False):
+        """
+        Process EMG data.
+        Parameters
+        ----------
+        data : numpy.ndarray
+            EMG data.
+        mvc_list : list
+            List of MVC values.
+        band_pass_filter : bool
+            Apply band pass filter.
+        low_pass_filter : bool
+            Apply low pass filter.
+        moving_average : bool
+            Apply moving average.
+        centering : bool
+            Apply centering.
+        absolute_value : bool
+            Apply absolute value.
+        normalization : bool
+            Apply normalization.
+        Returns
+        -------
+        numpy.ndarray
+            Processed EMG data.
+        """
+        data_proc = data
+        if band_pass_filter:
+            data_proc = self._butter_bandpass_filter(data, self.bpf_lcut, self.bpf_hcut, self.data_rate)
+        if centering:
+            data_proc = self.center(data_proc)
+        if absolute_value:
+            data_proc = np.abs(data_proc)
+        if low_pass_filter and moving_average:
+            raise RuntimeError("Please choose between low-pass filter and moving average.")
+        if low_pass_filter:
+            data_proc = self.butter_lowpass_filter(
+                data_proc, self.lpf_lcut, self.data_rate, order=self.lp_butter_order
+            )
+        else:
+            w = np.repeat(1, self.moving_average_windows) / self.moving_average_windows
+            empty_ma = np.ndarray((data.shape[0], data.shape[1]))
+            data_proc = self._moving_average(data_proc, w, empty_ma)
+
+        if normalization:
+            data_proc = self.normalize_emg(data_proc, mvc_list)
+
+        return data_proc
+
 
 class RealTimeProcessing(GenericProcessing):
-    def __init__(self):
+    def __init__(self, data_rate: int = 2000, processing_windows: int = 2000, moving_average_window: int = 200):
         """
         Initialize the class for real time processing.
         """
-        self.emg_rate = 2000
-        self.emg_win = 2000
-        self.ma_win = 200
         super().__init__()
+        self.data_rate = data_rate
+        self.processing_window = processing_windows
+        self.ma_win = moving_average_window
+        self.raw_data_buffer = []
+        self.processed_data_buffer = []
 
-    def process_emg(self,
-                    raw_emg: np.ndarray,
-                    emg_proc: np.ndarray,
-                    emg_tmp: np.ndarray,
-                    mvc_list: Union[list, tuple],
-                    norm_emg: bool = True,
-                    lpf: bool = False):
+    def process_emg(
+        self,
+        emg_data: np.ndarray,
+        mvc_list: Union[list, tuple] = None,
+        band_pass_filter = True,
+        low_pass_filter = False,
+        moving_average = True,
+        centering = True,
+        absolute_value = True,
+        normalization = False,
+        ):
         """
         Process EMG data in real-time.
         Parameters
         ----------
-        raw_emg : numpy.ndarray
-            Raw EMG data. (nb_emg, emg_win * emg_sample)
-        emg_proc : numpy.ndarray
-            Last processed EMG data. (nb_emg, emg_win)
-        emg_tmp : numpy.ndarray
+        emg_data : numpy.ndarray
             Temporary EMG data (nb_emg, emg_sample).
         mvc_list : list
             MVC values.
-        norm_emg : bool
-            Normalize EMG data.
-        lpf : bool
-            Apply low-pass filter.
         Returns
         -------
         tuple
             raw and processed EMG data.
 
         """
-        if self.ma_win > self.emg_win:
-            raise RuntimeError(f"Moving average windows ({self.ma_win}) higher than emg windows ({self.emg_win}).")
-        emg_sample = emg_tmp.shape[1]
+        if self.ma_win > self.processing_window:
+            raise RuntimeError(f"Moving average windows ({self.ma_win}) higher than emg windows ({self.processing_window}).")
+        emg_sample = emg_data.shape[1]
 
-        if norm_emg is True:
+        if normalization:
+            if not mvc_list:
+                raise RuntimeError("Please give a list of mvc to normalize the emg signal.")
             if isinstance(mvc_list, np.ndarray) is True:
                 if len(mvc_list.shape) == 1:
                     quot = mvc_list.reshape(-1, 1)
@@ -146,70 +222,56 @@ class RealTimeProcessing(GenericProcessing):
         else:
             quot = [1]
 
-        if len(raw_emg) == 0:
-            emg_proc = np.zeros((emg_tmp.shape[0], 1))
-            raw_emg = emg_tmp
+        if len(self.raw_data_buffer) == 0:
+            self.processed_data_buffer = np.zeros((emg_data.shape[0], 1))
+            self.raw_data_buffer = emg_data
 
-        elif raw_emg.shape[1] < self.emg_win:
-            emg_proc = np.zeros((emg_tmp.shape[0], self.emg_win))
-            raw_emg = np.append(raw_emg, emg_tmp, axis=1)
+        elif self.raw_data_buffer.shape[1] < self.processing_window:
+            self.raw_data_buffer = np.append(self.raw_data_buffer, emg_data, axis=1)
+            self.processed_data_buffer = np.zeros((emg_data.shape[0], self.raw_data_buffer.shape[1]))
 
         else:
-            raw_emg = np.append(raw_emg[:, -self.emg_win + emg_sample:], emg_tmp, axis=1)
-            emg_proc_tmp = abs(self.center(self._butter_bandpass_filter(raw_emg,
-                                                                        self.bpf_lcut,
-                                                                        self.bpf_hcut,
-                                                                        self.emg_rate
-                                                                        )
-                                           )
-                               )
-
-            if lpf is True:
-                emg_lpf_tmp = self.butter_lowpass_filter(emg_proc_tmp, self.lpf_lcut, self.emg_rate, order=self.lp_butter_order)
+            self.raw_data_buffer = np.append(self.raw_data_buffer[:, -self.processing_window + emg_sample :], emg_data, axis=1)
+            emg_proc_tmp =  self._process_emg(self.raw_data_buffer, band_pass_filter=band_pass_filter,
+                                                     centering=centering,
+                                                     absolute_value=absolute_value,
+                                                     low_pass_filter=False,
+                                                     moving_average=False,
+                                                     normalization=False
+                                                     )
+            if low_pass_filter and moving_average:
+                raise RuntimeError("Please choose between low-pass filter and moving average.")
+            if low_pass_filter:
+                emg_lpf_tmp = self.butter_lowpass_filter(
+                    emg_proc_tmp, self.lpf_lcut, self.data_rate, order=self.lp_butter_order
+                )
                 emg_lpf_tmp = emg_lpf_tmp[:, ::emg_sample]
-                emg_proc = np.append(emg_proc[:, emg_sample:], emg_lpf_tmp[:, -emg_sample:], axis=1)
-
+                self.processed_data_buffer = np.append(self.processed_data_buffer[:, emg_sample:], emg_lpf_tmp[:, -emg_sample:], axis=1)
             else:
                 average = np.median(emg_proc_tmp[:, -self.ma_win:], axis=1).reshape(-1, 1)
-                emg_proc = np.append(emg_proc[:, 1:], average / quot, axis=1)
-        return raw_emg, emg_proc
+                self.processed_data_buffer = np.append(self.processed_data_buffer[:, 1:], average / quot, axis=1)
+        return self.processed_data_buffer
 
-    @staticmethod
-    def process_imu(
-            im_proc,
-            raw_im,
-            im_tmp,
-            im_win,
-            im_sample,
-            ma_win,
-            accel=False,
-            squared=False,
-            norm_min_bound=None,
-            norm_max_bound=None,
+    def process_imu(self,
+        im_data: np.ndarray,
+        accel: bool =False,
+        squared: bool =False,
+        norm_min_bound:int=None,
+        norm_max_bound: int=None,
     ):
         """
         Process IMU data in real-time.
         Parameters
         ----------
-        im_proc : numpy.ndarray
-            Last processed IMU data. (nb_imu, im_win)
-        raw_im : numpy.ndarray
-            Raw IMU data. (nb_imu, im_win * im_sample)
-        im_tmp : numpy.ndarray
+        im_data : numpy.ndarray
             Temporary IMU data (nb_imu, im_sample).
-        im_win : int
-            IMU window size.
-        im_sample : int
-            IMU sample size.
-        ma_win : int
-            Moving average window size.
         accel : bool
             If current data is acceleration data to adapt the processing.
         squared : bool
             Apply squared.
-        norm_min_bound : tuple
+        norm_min_bound : int
             Normalization minimum bound.
-        norm_max_bound : tuple
+        norm_max_bound : int
             Normalization maximum bound.
 
         Returns
@@ -218,24 +280,24 @@ class RealTimeProcessing(GenericProcessing):
             raw and processed IMU data.
         """
 
-        if len(raw_im) == 0:
+        if len(self.raw_data_buffer) == 0:
             if squared is not True:
-                im_proc = np.zeros((im_tmp.shape[0], im_tmp.shape[1], 1))
+                self.processed_data_buffer = np.zeros((im_data.shape[0], im_data.shape[1], 1))
             else:
-                im_proc = np.zeros((im_tmp.shape[0], 1))
-            raw_im = im_tmp
+                self.processed_data_buffer = np.zeros((im_data.shape[0], 1))
+            self.raw_data_buffer = im_data
 
-        elif raw_im.shape[2] < im_win:
+        elif self.raw_data_buffer.shape[2] < self.processing_window:
+            self.raw_data_buffer = np.append(self.raw_data_buffer, im_data, axis=2)
             if squared is not True:
-                im_proc = np.zeros((im_tmp.shape[0], im_tmp.shape[1], im_win))
+                self.processed_data_buffer = np.zeros((im_data.shape[0], im_data.shape[1], self.raw_data_buffer.shape[2]))
             else:
-                im_proc = np.zeros((im_tmp.shape[0], im_win))
-            raw_im = np.append(raw_im, im_tmp, axis=2)
+                self.processed_data_buffer = np.zeros((im_data.shape[0], self.raw_data_buffer.shape[2]))
 
         else:
-            raw_im = np.append(raw_im[:, :, -im_win + im_sample:], im_tmp, axis=2)
-            im_proc_tmp = raw_im
-            average = np.mean(im_proc_tmp[:, :, -ma_win:], axis=2).reshape(-1, 3, 1)
+            self.raw_data_buffer = np.append(self.raw_data_buffer[:, :, -self.processing_window + im_data.shape[2] :], im_data, axis=2)
+            im_proc_tmp = self.raw_data_buffer
+            average = np.mean(im_proc_tmp[:, :, -self.processing_window:], axis=2).reshape(-1, 3, 1)
             if squared:
                 if accel:
                     average = abs(np.linalg.norm(average, axis=1) - 9.81)
@@ -244,36 +306,37 @@ class RealTimeProcessing(GenericProcessing):
 
             if len(average.shape) == 3:
                 if norm_min_bound or norm_max_bound:
-                    for i in range(raw_im.shape[0]):
-                        for j in range(raw_im.shape[1]):
+                    for i in range(self.raw_data_buffer .shape[0]):
+                        for j in range(self.raw_data_buffer .shape[1]):
                             if average[i, j, :] < 0:
                                 average[i, j, :] = average[i, j, :] / abs(norm_min_bound)
                             elif average[i, j, :] >= 0:
                                 average[i, j, :] = average[i, j, :] / norm_max_bound
-                im_proc = np.append(im_proc[:, :, 1:], average, axis=2)
+                self.processed_data_buffer = np.append(self.processed_data_buffer[:, :, 1:], average, axis=2)
 
             else:
                 if norm_min_bound or norm_max_bound:
-                    for i in range(raw_im.shape[0]):
-                        for j in range(raw_im.shape[1]):
+                    for i in range(self.raw_data_buffer.shape[0]):
+                        for j in range(self.raw_data_buffer.shape[1]):
                             if average[i, :] < 0:
                                 average[i, :] = average[i, :] / abs(norm_min_bound)
                             elif average[i, :] >= 0:
                                 average[i, :] = average[i, :] / norm_max_bound
-                im_proc = np.append(im_proc[:, 1:], average, axis=1)
+                self.processed_data_buffer = np.append(self.processed_data_buffer[:, 1:], average, axis=1)
 
-        return raw_im, im_proc
+        return self.processed_data_buffer
 
     @staticmethod
-    def get_peaks(new_sample: np.ndarray,
-                  signal: np.ndarray,
-                  signal_proc: np.ndarray,
-                  threshold: float,
-                  chanel_idx: Union[int, list] = None,
-                  nb_min_frame: float = 2000,
-                  is_one = None,
-                  min_peaks_interval = None
-                  ):
+    def get_peaks(
+        new_sample: np.ndarray,
+        signal: np.ndarray,
+        signal_proc: np.ndarray,
+        threshold: float,
+        chanel_idx: Union[int, list] = None,
+        nb_min_frame: float = 2000,
+        is_one=None,
+        min_peaks_interval=None,
+    ):
         """
         Allow to get the number of peaks for an analog signal (to get cadence from treadmill for instance).
         Parameters
@@ -318,8 +381,8 @@ class RealTimeProcessing(GenericProcessing):
             nb_peaks = None
 
         else:
-            signal = np.append(signal[:, new_sample.shape[1]:], new_sample, axis=1)
-            signal_proc = np.append(signal_proc[:, new_sample.shape[1]:], sample_proc, axis=1)
+            signal = np.append(signal[:, new_sample.shape[1] :], new_sample, axis=1)
+            signal_proc = np.append(signal_proc[:, new_sample.shape[1] :], sample_proc, axis=1)
 
         if chanel_idx:
             signal = signal[chanel_idx, :]
@@ -342,70 +405,51 @@ class RealTimeProcessing(GenericProcessing):
         return signal
 
     @staticmethod
-    def custom_processing(funct, raw_data, data_proc, data_tmp, *args, **kwargs):
-        return funct(raw_data, data_proc, data_tmp, *args, **kwargs)
+    def custom_processing(funct, data_tmp, **kwargs):
+        return funct(data_tmp, **kwargs)
 
 
 class OfflineProcessing(GenericProcessing):
-    def __init__(self):
+    def __init__(self, data_rate: float = 2000, processing_window: int = 2000, moving_average_windows: int = 200):
         """
         Offline processing.
         """
-        super().__init__()
+        super(OfflineProcessing, self).__init__()
+        self.data_rate = data_rate
+        self.processing_window = processing_window
+        self.moving_average_windows = moving_average_windows
 
-    def process_emg(self, data, frequency, pyomeca=False, ma=False):
+    @staticmethod
+    def custom_processing(funct, raw_data, **kwargs):
+        return funct(raw_data, **kwargs)
+
+    def process_emg(self, data: np.ndarray, mvc_list: list)-> np.ndarray:
         """
         Process EMG data.
         Parameters
         ----------
-        data : numpy.ndarray
-            EMG data.
-        frequency : int
-            EMG data frequency.
-        pyomeca : bool
-            If true, use low pass filter from pyomeca.
-        ma : bool
-            If true, apply moving average.
+        data : np.ndarray
+            Raw EMG data.
+        mvc_list: list
+            List of MVC for each muscle.
+
         Returns
-        -------
-        numpy.ndarray
+        np.ndarray
             Processed EMG data.
-        """
-        if pyomeca is True:
-            if pyomeca_module is False:
-                raise RuntimeError("Pyomeca module not found.")
-            if ma is True:
-                raise RuntimeError("Moving average not available with pyomeca.")
-            emg = Analogs(data)
-            emg_processed = (
-                emg.meca.band_pass(order=self.bp_butter_order, cutoff=[self.bpf_lcut, self.bpf_hcut], freq=frequency)
-                    .meca.abs()
-                    .meca.low_pass(order=self.lp_butter_order, cutoff=self.lpf_lcut, freq=frequency)
-            )
-            emg_processed = emg_processed.values
-
-        else:
-            emg_processed = abs(self.center(self._butter_bandpass_filter(data, self.bpf_lcut, self.bpf_hcut, frequency)))
-            if ma is True:
-                w = np.repeat(1, self.ma_win) / self.ma_win
-                empty_ma = np.ndarray((data.shape[0], data.shape[1]))
-                emg_processed = self._moving_average(emg_processed, w, empty_ma)
-            else:
-                emg_processed = self.butter_lowpass_filter(emg_processed, self.lpf_lcut, frequency, order=self.lp_butter_order)
-        return emg_processed
+        -------
+    """
+        return self._process_emg(data, mvc_list)
 
     @staticmethod
-    def custom_processing(funct, raw_data, *args, **kwargs):
-        return funct(raw_data, *args, **kwargs)
-
-    @staticmethod
-    def compute_mvc(nb_muscles: int,
-                    mvc_trials: np.ndarray,
-                    window_size: int,
-                    mvc_list_max: np.ndarray,
-                    tmp_file: str = None,
-                    output_file: str = None,
-                    save: bool = False):
+    def compute_mvc(
+        nb_muscles: int,
+        mvc_trials: np.ndarray,
+        window_size: int,
+        mvc_list_max: np.ndarray,
+        tmp_file: str = None,
+        output_file: str = None,
+        save: bool = False,
+    ):
         """
         Compute MVC from several mvc_trials.
 
@@ -448,5 +492,6 @@ class OfflineProcessing(GenericProcessing):
 
         if save:
             sio.savemat(output_file, mat_content)
-        os.remove(tmp_file)
+        if tmp_file:
+            os.remove(tmp_file)
         return mvc_list_max
